linabe/requirements.txt
Django>=3.1.2,<3.2.0
djangorestframework>=3.12.1,<3.13.0
django-cors-headers>=3.5.0,<3.6.0
Pillow>=7.2.0,<7.3.0
flake8>=3.8.3,<3.9.0
mysqlclient>=2.0.1,<2.2.0
django-ckeditor>=6.0.0,<6.1.0
django-filter>=2.4.0,<2.5.0
djangorestframework-simplejwt
gunicorn>=20.0.4,<20.1.0
cx-Oracle>=8.1.0,<8.2.0
drf-dx-datagrid>=0.4.1,<1.0.1
django-ipware>=3.0.2,<3.1.0
weasyprint==52.5

# Celery and related tools
celery>=5.2.0,<5.3.0
watchfiles
flower>=1.0.0,<2.0.0
redis>=4.0.0,<5.0.0
daphne>=3.0.0,<4.0.0
channels>=3.0.0,<4.0.0
channels-redis>=3.3.1,<4.0.0
django-celery-beat>=2.2.1,<2.3.0